move DecisionViewFooter outside SubmitDecisionView
