+++ client/app/queue/SubmitDecisionView.jsx
@@ -3,7 +3,6 @@ import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import _ from 'lodash';
 import { css } from 'glamor';
-import { withRouter } from 'react-router-dom';
 
 import AppSegment from '@department-of-veterans-affairs/caseflow-frontend-toolkit/components/AppSegment';
 import RadioField from '../components/RadioField';
@@ -11,7 +10,6 @@ import Checkbox from '../components/Checkbox';
 import TextField from '../components/TextField';
 import TextareaField from '../components/TextareaField';
 import Button from '../components/Button';
-import DecisionViewFooter from './components/DecisionViewFooter';
 
 import { fullWidth } from './constants';
 
@@ -43,23 +41,6 @@ class SubmitDecisionView extends React.PureComponent {
   }
 
   render = () => {
-    const {
-      appeal: { attributes: appeal },
-      history,
-      vacolsId
-    } = this.props;
-    const footerButtons = [{
-      displayText: `Go back to ${appeal.veteran_full_name}`,
-      callback: () => {
-        history.push(`/tasks/${vacolsId}`);
-        window.scrollTo(0, 0);
-      },
-      classNames: ['cf-btn-link']
-    }, {
-      displayText: 'Submit',
-      callback: () => console.warn('data submitted'),
-      classNames: ['cf-right-side']
-    }];
     const omoTypes = [{
       displayText: 'VHA - OMO',
       value: 'omo'
@@ -68,54 +49,51 @@ class SubmitDecisionView extends React.PureComponent {
       value: 'ime'
     }];
 
-    return <React.Fragment>
-      <AppSegment filledBackground>
-        <h1 className="cf-push-left" {...css(fullWidth, smallBottomMargin)}>
-          Submit OMO
-        </h1>
-        <p className="cf-lead-paragraph" {...subHeadStyling}>
-          Review and complete the following details to check this task out for judge review.
-        </p>
-        <hr />
-        <RadioField
-          name="omo_type"
-          label="OMO type:"
-          onChange={(omoType) => this.setState({ omo_type: omoType })}
-          value={this.state.omo_type}
-          vertical
-          required
-          options={omoTypes}
-          styling={radioFieldStyling}
-        />
-        <Checkbox
-          name="overtime"
-          label="This work product is overtime"
-          onChange={(overtime) => this.setState({ overtime })}
-          value={this.state.overtime}
-          styling={css(smallBottomMargin, checkboxStyling)}
-        />
-        <TextField
-          name="Document ID:"
-          required
-          onChange={(documentId) => this.setState({ document_id: documentId })}
-          value={this.state.document_id}
-        />
-        <span>Check out to:</span><br />
-        <span>Nick Kroes</span>
-        <Button
-          classNames={['cf-btn-link']}
-          onClick={_.noop}>
-          Select another judge
-        </Button>
-        <TextareaField
-          name="Notes:"
-          value={this.state.notes}
-          onChange={(notes) => this.setState({ notes })}
-          styling={textAreaStyling}
-        />
-      </AppSegment>
-      <DecisionViewFooter buttons={footerButtons} />
-    </React.Fragment>;
+    return <AppSegment filledBackground>
+      <h1 className="cf-push-left" {...css(fullWidth, smallBottomMargin)}>
+        Submit OMO
+      </h1>
+      <p className="cf-lead-paragraph" {...subHeadStyling}>
+        Review and complete the following details to check this task out for judge review.
+      </p>
+      <hr />
+      <RadioField
+        name="omo_type"
+        label="OMO type:"
+        onChange={(omoType) => this.setState({ omo_type: omoType })}
+        value={this.state.omo_type}
+        vertical
+        required
+        options={omoTypes}
+        styling={radioFieldStyling}
+      />
+      <Checkbox
+        name="overtime"
+        label="This work product is overtime"
+        onChange={(overtime) => this.setState({ overtime })}
+        value={this.state.overtime}
+        styling={css(smallBottomMargin, checkboxStyling)}
+      />
+      <TextField
+        name="Document ID:"
+        required
+        onChange={(documentId) => this.setState({ document_id: documentId })}
+        value={this.state.document_id}
+      />
+      <span>Check out to:</span><br />
+      <span>Nick Kroes</span>
+      <Button
+        classNames={['cf-btn-link']}
+        onClick={_.noop}>
+        Select another judge
+      </Button>
+      <TextareaField
+        name="Notes:"
+        value={this.state.notes}
+        onChange={(notes) => this.setState({ notes })}
+        styling={textAreaStyling}
+      />
+    </AppSegment>;
   };
 }
 
@@ -127,4 +105,4 @@ const mapStateToProps = (state, ownProps) => ({
   appeal: state.queue.loadedQueue.appeals[ownProps.vacolsId]
 });
 
-export default withRouter(connect(mapStateToProps)(SubmitDecisionView));
+export default connect(mapStateToProps)(SubmitDecisionView);
